// src/types.ts
export interface Card {
  id: number;
  card_number: number;
  name: string;
  group_name: string;
  element: string;
  legend: string;
  general_meaning: string;
  work: string;
  finance: string;
  love: string;
  health: string;
  merit: string;
  visual_key: string;
  deep_meaning: string;
  occult: string;
  image_url: string;
}